Move `event` module under the `server` module
