+++ crates/hyperqueue/src/client/commands/server.rs
@@ -1,10 +1,12 @@
-use crate::client::commands::stats::print_server_stats;
 use crate::client::globalsettings::GlobalSettings;
 use crate::client::server::client_stop_server;
 use crate::common::timeutils::ArgDuration;
+use crate::rpc_call;
 use crate::server::bootstrap::{
     get_client_connection, init_hq_server, print_server_info, ServerConfig,
 };
+use crate::transfer::connection::ClientConnection;
+use crate::transfer::messages::{FromClientMessage, StatsResponse, ToClientMessage};
 use clap::Parser;
 use std::path::PathBuf;
 
@@ -105,4 +107,19 @@ async fn command_server_info(
     } else {
         print_server_info(gsettings).await
     }
+}
+
+async fn print_server_stats(
+    gsettings: &GlobalSettings,
+    connection: &mut ClientConnection,
+) -> anyhow::Result<()> {
+    let response: StatsResponse = rpc_call!(
+        connection,
+        FromClientMessage::Stats,
+        ToClientMessage::StatsResponse(r) => r
+    )
+    .await?;
+
+    gsettings.printer().print_server_stats(response);
+    Ok(())
 }
