+++ crates/hyperqueue/src/dashboard/data/mod.rs
@@ -5,8 +5,8 @@ use tako::messages::common::WorkerConfiguration;
 use tako::messages::gateway::MonitoringEventRequest;
 use tako::messages::worker::WorkerOverview;
 
-use crate::event::events::MonitoringEventPayload;
-use crate::event::MonitoringEvent;
+use crate::server::event::events::MonitoringEventPayload;
+use crate::server::event::MonitoringEvent;
 use crate::transfer::connection::ClientConnection;
 use crate::transfer::messages::{FromClientMessage, ToClientMessage};
 use crate::{rpc_call, WorkerId};
