+++ crates/hyperqueue/src/server/mod.rs
@@ -1,6 +1,7 @@
 pub mod autoalloc;
 pub mod bootstrap;
 pub mod client;
+pub mod event;
 pub mod job;
 pub mod rpc;
 pub mod state;
